refactor: split wandb metrics and publication
